IGVF-60-report-view
Rearranged the code to split off the collection table view code into its own file, and made a new file for a standard checkbox.
