--- components/table-view.js
@@ -1,7 +1,0 @@
-import { DataGridContainer } from "./data-grid"
-
-const TableView = ({ children }) => {
-  return <DataGridContainer>{children}</DataGridContainer>
-}
-
-export default TableView
